fix display of thread id in debug log

+++ debuglog.cpp
@@ -160,9 +160,9 @@ wxString DebugLog::Write(const wxString& str)
         wxDateTime now = wxDateTime::UNow();
         wxTimeSpan deltaTime = now - m_lastWriteTime;
         m_lastWriteTime = now;
-        wxString outputLine = wxString::Format("%s %s %d %s", now.Format("%H:%M:%S.%l"),
+        wxString outputLine = wxString::Format("%s %s %lu %s", now.Format("%H:%M:%S.%l"),
                                                               deltaTime.Format("%S.%l"),
-                                                              wxThread::GetCurrentId(),
+                                                              (unsigned long) wxThread::GetCurrentId(),
                                                               str);
